requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
blpapi>=3.20.0
supabase>=2.0.0
pyodbc>=4.0.0
python-dateutil>=2.8.0
